Change views, forms and design elements to take advantage of enum
Yeah, this should be in another branch but what do you want from me?
I'm not perfect, bub. Most changes are on the schedule _form partial
to keep it from dying with the new group enum data. Jquery UI buttonset
is also used to clean up the look a little. Change the Jq-UI CSS to
fix some silly font sizes and borders too.